+++ apps/fluent-tester/src/FluentTester/TestComponents/Common/styles.ts
@@ -96,7 +96,10 @@ export const fluentTesterStyles = StyleSheet.create({
   },
 
   testList: {
-    width: 160,
+    minWidth: 160,
+    flexShrink: 0,
+    flexGrow: 0,
+    flexDirection: 'row',
   },
 
   testListContainerStyle: {