+++ packages/components/Persona/src/Persona.tokens.coin.ts
@@ -1,4 +1,4 @@
-import { ViewProps } from 'react-native';
+import type { IViewProps } from '@fluentui-react-native/adapters';
 import { styleFunction } from '@uifabricshared/foundation-tokens';
 import { ITheme } from '@uifabricshared/theming-ramp';
 import { buildRootStyles as buildPersonaCoinRootStyle, IPersonaCoinTokens } from '@fluentui-react-native/persona-coin';
@@ -9,18 +9,18 @@ const _coinStyleKeys: (keyof IPersonaTokens)[] = [
   'size',
   'horizontalIconAlignment',
   'verticalIconAlignment',
-  'coinBackgroundColor'
+  'coinBackgroundColor',
 ];
 
-function _buildCoinStyle(tokenProps: IPersonaTokens, theme: ITheme): ViewProps {
+function _buildCoinStyle(tokenProps: IPersonaTokens, theme: ITheme): IViewProps {
   const { coinBackgroundColor, ...rest } = tokenProps;
 
   const personaCoinTokens: IPersonaCoinTokens = {
     ...rest,
-    backgroundColor: coinBackgroundColor
+    backgroundColor: coinBackgroundColor,
   };
 
   return buildPersonaCoinRootStyle(personaCoinTokens, theme);
 }
 
-export const buildCoinStyle = styleFunction<ViewProps, IPersonaTokens, ITheme>(_buildCoinStyle, _coinStyleKeys);
+export const buildCoinStyle = styleFunction<IViewProps, IPersonaTokens, ITheme>(_buildCoinStyle, _coinStyleKeys);