+++ packages/theming/apple-theme/src/createAppleTheme.ios.ts
@@ -1,7 +1,17 @@
 import { ThemeReference } from '@fluentui-react-native/theme';
-import { createDefaultTheme } from '@fluentui-react-native/default-theme';
+import { Theme } from '@fluentui-react-native/theme-types';
+import { Appearance } from 'react-native';
+import { BaseAppleDarkThemeIOS, BaseAppleLightThemeIOS } from './appleTheme.ios';
 
 export function createAppleTheme(): ThemeReference {
-  console.warn('iOS Theme not currently implemented, using default theme');
-  return new ThemeReference(createDefaultTheme());
+  const themeRef = new ThemeReference({} as Theme, () => {
+    const current = Appearance.getColorScheme();
+    return current === 'light' ? BaseAppleLightThemeIOS : BaseAppleDarkThemeIOS;
+  });
+
+  Appearance.addChangeListener(() => {
+    themeRef.invalidate();
+  });
+
+  return themeRef;
 }