+++ apps/fluent-tester/CHANGELOG.md
@@ -1,8 +1,22 @@
 # Change Log - @fluentui-react-native/tester
 
-This log was last generated on Mon, 01 Mar 2021 22:53:12 GMT and should not be manually modified.
+This log was last generated on Sat, 06 Mar 2021 00:20:05 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.17.0
+
+Sat, 06 Mar 2021 00:20:05 GMT
+
+### Minor changes
+
+- Base theme on macOS changed to apple theme (saadnajmi2@gmail.com)
+- Fixed a lot of layout bugs (saadnajmi2@gmail.com)
+- Update to react-native 0.63 (30809111+acoates-ms@users.noreply.github.com)
+
+### Patches
+
+- Add Experimental Button to iOS tests (saadnajmi2@gmail.com)
 
 ## 0.16.3